+++ src/components/CartPageComponents/StepperInput/StepperInput.tsx
@@ -1,6 +1,6 @@
 import * as Styled from './StepperInput.styles.tsx';
 import { Item } from '../../../types/CartList.ts';
-import useCartItemOperations from '../../../hooks/useCartItemOperations.ts';
+import useCartItemOperations from '../../../hooks/cartItemOperations/useCartItemOperations.ts';
 
 type StepperInputProps = {
   initialValue: number;
